refactor: Move ELF-specific allocation of internal symbols
